Move ObserveElementProperty for ROC to CollectionUtilities
